+++ app/db/models/tenant.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 from typing import Optional
-from sqlalchemy import String, BigInt, TIMESTAMP, Enum, func
+from sqlalchemy import String, BigInteger, TIMESTAMP, Enum, func
 from sqlalchemy.orm import Mapped, mapped_column
 from app.db.session import Base
 import enum
@@ -13,7 +13,7 @@ class TenantStatus(str, enum.Enum):
 class Tenant(Base):
     __tablename__ = "tenants"
 
-    id: Mapped[int] = mapped_column(BigInt, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
     tenant_uuid: Mapped[str] = mapped_column(String(36), unique=True, nullable=False, index=True)
     name: Mapped[str] = mapped_column(String(150), nullable=False)
     slug: Mapped[str] = mapped_column(String(100), unique=True, nullable=False)
@@ -22,8 +22,8 @@ class Tenant(Base):
     )
     default_model: Mapped[Optional[str]] = mapped_column(String(150), nullable=True)
     default_embedding_model: Mapped[Optional[str]] = mapped_column(String(150), nullable=True)
-    monthly_quota_tokens: Mapped[Optional[int]] = mapped_column(BigInt, nullable=True)
-    monthly_quota_requests: Mapped[Optional[int]] = mapped_column(BigInt, nullable=True)
+    monthly_quota_tokens: Mapped[Optional[int]] = mapped_column(BigInteger, nullable=True)
+    monthly_quota_requests: Mapped[Optional[int]] = mapped_column(BigInteger, nullable=True)
     
     created_at: Mapped[datetime] = mapped_column(
         TIMESTAMP, nullable=False, server_default=func.now()
